requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.0
matplotlib==3.7.2
flask==3.0.0
werkzeug==3.0.1
